termId uten ordklasse, heltall indikerer antall identiske termer i db
