+++ api/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "author": "Isak Kyrre Lichtwarck Bjugn",
   "license": "ISC",
@@ -18,5 +18,8 @@
     "express": "4.18.2",
     "google-auth-library": "9.2.0",
     "google-spreadsheet": "4.1.0"
+  },
+  "devDependencies": {
+    "jest": "29.7.0"
   }
 }
